+++ Sources/StreamChat/Repositories/PollsRepository.swift
@@ -205,13 +205,13 @@ class PollsRepository {
         pollId: String,
         completion: ((Error?) -> Void)? = nil
     ) {
-        database.write { session in
-            _ = try? session.deletePoll(pollId: pollId)
-        } completion: { [weak self] _ in
-            // No need to check for error, since the poll can also be deleted independently.
-            self?.apiClient.request(endpoint: .deletePoll(pollId: pollId)) {
-                completion?($0.error)
+        apiClient.request(endpoint: .deletePoll(pollId: pollId)) { [weak self] in
+            if $0.error == nil {
+                self?.database.write { session in
+                    _ = try? session.deletePoll(pollId: pollId)
+                }
             }
+            completion?($0.error)
         }
     }